refactor: move MerchantDetails, PayPalOrder and PayPalPayment to Models directory
